+++ demos/demos/common/utils/base32.py
@@ -75,7 +75,7 @@ def random(length, crypto=True):
 	length of randomly encoded string."""
 	
 	bits = length * 5
-	bytes = math.ceil(bits / 8)
+	bytes = int(math.ceil(bits / 8.0))
 	shift_bits = (8 * bytes) - bits
 	
 	if crypto:
